Assing the correct security groups to load balancer and instances. Test playbook for the master access through the external load balancer
Security groups have been assigned to external load balancer and instances to allow the access from the Internet to the masters on ports 443 and 8444.

ssh config file uses FQDN instead of alias for bastion, and domain name to access private subnet EC2 instances.

A playbook to deploy httpd servers on the masters, so the access to this machines through the load balancer can be tested, have been created

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>