Add load balancer in front of infra nodes.  Add script to create inventory file
http-test.yaml playbook runs against nodes with specific variable definition in the inventory file

A variable is defined in terraform file to use it for naming porpuses in the cluster.

The load balancer sitting in front of the infra nodes is added, and the corresponding security groups to the infra nodes.

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>