Add DNS internal and external zones, along some records in them.
Only one aws provider is used, the DNS zones are created as subzones of rhcee.support, one external to resolve the names of services accessable from the internet like the master public load balancer or the applications domain.  The other zone is internal to resolve DNS names inside the same cluster.

Bastion name in the inventory is based on the output from terraform.

Bastion name in ssh.cfg file uses the DNS domain ocpext.rhcee.support.  This also should be dynamically created like the inventory name.