Complete the definition of EC2 instances for the cluster nodes.  Add docker test ansible playbook
Added variables to define the instance type associated with masters and nodes

Added variables containing the user data for masters and rest of nodes, the user data contents define cloud config directives to set up disks and partitions, VG/LV needed by docker, etcd, etc.  Depending on the type of ec2 instance used the disk device use different naming conventions: /dev/nvme1n1 vs /dev/xvda

The node security group is applied to all nodes in the cluster, since this is a base security group with rules applying to all.

Additional disks are attached to masters and nodes to be used by docker and openshift as independent partitions, the root disks are resized according to the recommendations in the OpenShift documentation.

Added ingress rule to the sg-master security group so nodes can connect to masters on port 2379 and 2380, as documentation explains.

An ansible playbook has been added to verify that all nodes can run docker containers.

 resource "aws_vpc" "vpc" {
     cidr_block = "172.20.0.0/16"
@@ -364,6 +447,13 @@ resource "aws_security_group" "sg-master" {
     }

     ingress {
+        from_port = 2379
+        to_port = 2380
+        protocol = "tcp"
+        security_groups = [aws_security_group.sg-node.id]
+    }
+
+    ingress {
         from_port = 443
         to_port = 443
         protocol = "tcp"

   tags = {
         Name = "worker03"