Packages installed in bastion during prereqs-ocp.yaml playbook execution, and packages updated if var defined as true.
The create_inventory.sh script formats the output better and a new group [etcd:children] has been added to reflect the normal structure of an inventory file.

Some packages are installed in the bastion host: openshift-ansible; vim; screen

OS packages in the nodes will be updated if the variable "update_packages" is defined as true, by default it is false.

The DNS name for the internal load balancer infront of the masters is shown in the terraform output

The arn value of the S3 bucket for the registry is shown in the terraform output