Check for the correct deploy of the image registry and router during cluster deployment. Ansible 2.8 is used.
When the cluster is being deployed in us-east-1 AWS region the DNS subdomain distributed by the DHCP is changed to ec2.internal

When the cluster is being deployed in eu-north-1 AWS region a particular S3 endpoint is specified

Ansible 2.8 is now supported and a lot faster than previous 2.6

Default nodes instance type changed to t3.xlarge due to m4.xlarge not being available in many regions

Bastion uses the same instance type as the nodes.

Random string for S3 bucket names does not contain dashes (-) anymore, becuase it is not legal if it lands as the first character in the name.

IAM user names contain a random suffix of 5 characters to make sure they are unique within the organization.

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>

 #IAM users
+
+#Provides a source to create a random sufix string for the IAM user names
+resource "random_string" "sufix_name" {
+  length = 5
+  upper = false
+  special = false
+}
+
 #Admin user for aws OpenShift plugin
 resource "aws_iam_user" "iam-admin" {
-  name = "iam-admin"
+  name = "iam-admin-${random_string.sufix_name.result}"

    tags = {
     Name = "iam-admin"
@@ -950,7 +958,7 @@ resource "aws_iam_user_policy" "policy-admin" {

 #Registry user for S3 bucket access
 resource "aws_iam_user" "iam-registry" {
-  name = "iam-registry"
+  name = "iam-registry-${random_string.sufix_name.result}"

    tags = {
     Name = "iam-registry"